# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ADVISOR",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        },
        {
            "url": "https://advisor-api.barbacode.com"
        }
    ],
    "paths": {
        "/api/v1/user/register": {
            "post": {
                "tags": [
                    "register user"
                ],
                "summary": "API to register users",
                "description": "Send request to create users for login application",
                "operationId": "61f14b51d1fc8d87986d4e45686ce1b1",
                "requestBody": {
                    "description": "User object fields",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "national_id": {
                                        "type": "string",
                                        "maxLength": 8,
                                        "minLength": 8
                                    },
                                    "national_type": {
                                        "type": "string",
                                        "enum": [
                                            "dni",
                                            "pasaporte"
                                        ]
                                    },
                                    "grant_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "minLength": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User has been created"
                    },
                    "400": {
                        "description": "Bad request, please check request body"
                    }
                }
            }
        },
        "/api/v1/station/recovery-batch/{ruc}": {
            "get": {
                "tags": [
                    "search station by ruc"
                ],
                "summary": "Get stations recovered",
                "description": "Get stations recovery by batch from OSINERMING by searching ruc",
                "operationId": "8f9eacfd8503992ea851efe0679d95b0",
                "parameters": [
                    {
                        "name": "ruc",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "jwt"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stations was found like RUC parameter"
                    },
                    "401": {
                        "description": "Please check your token authorization"
                    }
                }
            }
        },
        "/api/v1/station/afiliar": {
            "post": {
                "tags": [
                    "To afiliate station"
                ],
                "summary": "To afiliate station",
                "description": "To afiliate stations in whosaler comapany",
                "operationId": "003b73fa5c922a1b5e24d184b15315c7",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "jwt"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request body to afiliate station",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ruc": {
                                        "type": "string",
                                        "maxLength": 8,
                                        "minLength": 8
                                    },
                                    "company_name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "address": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "ubigeo": {
                                        "properties": {
                                            "deparment": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "province": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "district": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "name": {
                                                    "type": "string"
                                                },
                                                "src": {
                                                    "type": "base64"
                                                },
                                                "type": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The station was afiliated"
                    },
                    "401": {
                        "description": "Please check your token authorization"
                    }
                }
            }
        },
        "/api/v1/station/afiliados": {
            "get": {
                "tags": [
                    "Get station afiliates"
                ],
                "summary": "Get alifiate stations",
                "description": "Get afiliate stations from whosaler station",
                "operationId": "b3b6486d251298a6083b0984f48fcaf0",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "jwt"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Afiliates stations obtain"
                    },
                    "401": {
                        "description": "Please check your token authorization"
                    }
                }
            }
        },
        "/api/v1/station/unset/{stationId}": {
            "delete": {
                "tags": [
                    "Unset afiliation of station"
                ],
                "summary": "Unset afiliation",
                "description": "Unset afiliaton of station",
                "operationId": "f30cda28555fffdac15c75f6241b5811",
                "parameters": [
                    {
                        "name": "stationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "jwt"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Station was removed of whosaler"
                    },
                    "401": {
                        "description": "Please check your token authorization"
                    },
                    "403": {
                        "description": "You dont have permission to unset this station"
                    }
                }
            }
        }
    }
}